Restructure perception/vision subpackages
