+++ src/robotics_utils/perception/pose_estimation/__init__.py
@@ -1,3 +1,6 @@
 """Import classes for pose estimation."""
 
 from .pose_estimate_averager import PoseEstimateAverager as PoseEstimateAverager
+from .visual_fiducials import FiducialCamera as FiducialCamera
+from .visual_fiducials import FiducialMarker as FiducialMarker
+from .visual_fiducials import FiducialSystem as FiducialSystem
