+++ src/robotics_utils/perception/sensors/realsense.py
@@ -12,7 +12,7 @@ from rich.table import Table
 from typing_extensions import Self
 
 from robotics_utils.perception.sensors.cameras import CameraIntrinsics, DepthCameraSpec, Resolution
-from robotics_utils.vision.images import DepthImage, RGBDImage, RGBImage
+from robotics_utils.perception.vision import DepthImage, RGBDImage, RGBImage
 
 if TYPE_CHECKING:
     from types import TracebackType
