+++ src/robotics_utils/perception/vision/bounding_box.py
@@ -8,10 +8,10 @@ from typing import TYPE_CHECKING
 import cv2
 import numpy as np
 
-from robotics_utils.vision.images import Image, PixelXY, RGBImage
+from robotics_utils.perception.vision.images import Image, PixelXY, RGBImage
 
 if TYPE_CHECKING:
-    from robotics_utils.vision.vision_utils import RGB
+    from robotics_utils.perception.vision.vision_utils import RGB
 
 
 @dataclass(frozen=True)
