+++ src/robotics_utils/ros/fiducial_tracker.py
@@ -11,8 +11,7 @@ from pose_estimation_msgs.msg import PoseEstimate
 from std_srvs.srv import Trigger, TriggerRequest, TriggerResponse
 
 from robotics_utils.kinematics import DEFAULT_FRAME
-from robotics_utils.perception.pose_estimation import PoseEstimateAverager
-from robotics_utils.perception.sensors.visual_fiducials import FiducialSystem
+from robotics_utils.perception.pose_estimation import FiducialSystem, PoseEstimateAverager
 from robotics_utils.ros.msg_conversion import pose_from_msg, pose_to_stamped_msg
 from robotics_utils.ros.params import get_ros_param
 from robotics_utils.ros.transform_manager import TransformManager
