+++ src/robotics_utils/visualization/__init__.py
@@ -2,3 +2,4 @@
 
 from .display_images import Displayable as Displayable
 from .display_images import display_image as display_image
+from .pointcloud_visualizer import PointcloudVisualizer as PointcloudVisualizer
