+++ tests/vision/test_vision_utils.py
@@ -4,7 +4,7 @@ import torch
 
 from robotics_utils.io.logging import log_info
 from robotics_utils.math.conversions import Gi_TO_B
-from robotics_utils.vision.vision_utils import determine_pytorch_device
+from robotics_utils.perception.vision import determine_pytorch_device
 
 
 def test_determine_pytorch_device() -> None:
